requirements.txt
python-dotenv == 0.21.0
pandas == 1.5.0
boto3 == 1.24.80
python-dotenv == 0.21.0
selenium == 4.4.3
webdriver-manager == 3.8.3